+++ packages/cli/src/commands/generate/forms/generate_forms_command.test.ts
@@ -51,14 +51,14 @@ void describe('generate forms command', () => {
       const modelsOutPath = './my-fake-models-path';
       const commandRunner = new TestCommandRunner(parser);
       await commandRunner.runCommand(
-        `forms --stack my_stack --modelsOutDir ${modelsOutPath}`
+        `forms --stack my_stack --models-out-dir ${modelsOutPath}`
       );
       assert.equal(
         generationMock.mock.calls[0].arguments[0].modelsOutDir,
         modelsOutPath
       );
     });
-    void it('uiOutDir path can be customized', async () => {
+    void it('ui-out-dir path can be customized', async () => {
       const credentialProvider = fromNodeProviderChain();
 
       const backendIdResolver = new BackendIdentifierResolver({
@@ -98,7 +98,7 @@ void describe('generate forms command', () => {
       const uiOutPath = './my-fake-ui-path';
       const commandRunner = new TestCommandRunner(parser);
       await commandRunner.runCommand(
-        `forms --stack my_stack --uiOutDir ${uiOutPath}`
+        `forms --stack my_stack --ui-out-dir ${uiOutPath}`
       );
       assert.equal(
         generationMock.mock.calls[0].arguments[0].uiOutDir,