+++ packages/cli/src/commands/generate/graphql-client-code/generate_graphql_client_code_command.test.ts
@@ -81,7 +81,7 @@ void describe('generate graphql-client-code command', () => {
 
   void it('generates and writes graphql client code for appID and branch', async () => {
     await commandRunner.runCommand(
-      'graphql-client-code --branch branch_name --appId app_id'
+      'graphql-client-code --branch branch_name --app-id app_id'
     );
     assert.equal(invokeGenerateApiCodeMock.mock.callCount(), 1);
     assert.deepEqual(invokeGenerateApiCodeMock.mock.calls[0].arguments[0], {
@@ -117,12 +117,12 @@ void describe('generate graphql-client-code command', () => {
   void it('shows available options in help output', async () => {
     const output = await commandRunner.runCommand('graphql-client-code --help');
     assert.match(output, /--stack/);
-    assert.match(output, /--appId/);
+    assert.match(output, /--app-id/);
     assert.match(output, /--branch/);
     assert.match(output, /--format/);
-    assert.match(output, /--statementTarget/);
-    assert.match(output, /--typeTarget/);
-    assert.match(output, /--modelTarget/);
+    assert.match(output, /--statement-target/);
+    assert.match(output, /--type-target/);
+    assert.match(output, /--model-target/);
     assert.match(output, /--out/);
     assert.match(output, /--all/);
   });
@@ -132,20 +132,20 @@ void describe('generate graphql-client-code command', () => {
       'graphql-client-code --help --all'
     );
     assert.match(output, /--stack/);
-    assert.match(output, /--appId/);
+    assert.match(output, /--app-id/);
     assert.match(output, /--branch/);
     assert.match(output, /--format/);
-    assert.match(output, /--statementTarget/);
-    assert.match(output, /--typeTarget/);
-    assert.match(output, /--modelTarget/);
+    assert.match(output, /--statement-target/);
+    assert.match(output, /--type-target/);
+    assert.match(output, /--model-target/);
     assert.match(output, /--out/);
     assert.match(output, /--all/);
-    assert.match(output, /--modelGenerateIndexRules/);
-    assert.match(output, /--modelEmitAuthProvider/);
-    assert.match(output, /--modelAddTimestampFields/);
-    assert.match(output, /--statementMaxDepth/);
-    assert.match(output, /--statementTypenameIntrospection/);
-    assert.match(output, /--typeMultipleSwiftFiles/);
+    assert.match(output, /--model-generate-index-rules/);
+    assert.match(output, /--model-emit-auth-provider/);
+    assert.match(output, /--model-add-timestamp-fields/);
+    assert.match(output, /--statement-max-depth/);
+    assert.match(output, /--statement-typename-introspection/);
+    assert.match(output, /--type-multiple-swift-files/);
   });
 
   void it('can be invoked explicitly with graphql-codegen format', async () => {
@@ -200,7 +200,7 @@ void describe('generate graphql-client-code command', () => {
 
   void it('passes in feature flags on modelgen', async () => {
     await commandRunner.runCommand(
-      'graphql-client-code --stack stack_name --format modelgen --modelGenerateIndexRules true --modelEmitAuthProvider true --modelGenerateModelsForLazyLoadAndCustomSelectionSet false'
+      'graphql-client-code --stack stack_name --format modelgen --model-generate-index-rules true --model-emit-auth-provider true --model-generate-models-for-lazy-load-and-custom-selection-set false'
     );
     assert.equal(invokeGenerateApiCodeMock.mock.callCount(), 1);
     assert.deepEqual(invokeGenerateApiCodeMock.mock.calls[0].arguments[0], {
@@ -220,7 +220,7 @@ void describe('generate graphql-client-code command', () => {
 
   void it('passes in feature flags on graphql-codegen', async () => {
     await commandRunner.runCommand(
-      'graphql-client-code --stack stack_name --format graphql-codegen --statementTarget typescript --statementMaxDepth 3 --statementTypenameIntrospection true'
+      'graphql-client-code --stack stack_name --format graphql-codegen --statement-target typescript --statement-max-depth 3 --statement-typename-introspection true'
     );
     assert.equal(invokeGenerateApiCodeMock.mock.callCount(), 1);
     assert.deepEqual(invokeGenerateApiCodeMock.mock.calls[0].arguments[0], {